Portability enhancement.

+++ darray.h
@@ -6,8 +6,8 @@
 #define DARRAY_INIT(T)                                                                                                 \
    typedef struct {                                                                                                    \
       T* node;                                                                                                         \
-      unsigned int size;                                                                                               \
-      unsigned int capacity;                                                                                           \
+      size_t size;                                                                                                     \
+      size_t capacity;                                                                                                 \
    } Darray_##T;                                                                                                       \
 static inline bool T##_push(Darray_##T* darray, T item) {                                                              \
    if (darray->capacity == 0) {                                                                                        \
@@ -23,18 +23,18 @@ static inline bool T##_push(Darray_##T* darray, T item) {                       
    darray->node[darray->size++] = item;                                                                                \
    return true;                                                                                                        \
 }                                                                                                                      \
-static inline void T##_pop(Darray_##T* darray, unsigned int count) {                                                   \
+static inline void T##_pop(Darray_##T* darray, size_t count) {                                                         \
    if (count > darray->size) darray->size = 0;                                                                         \
    else darray->size -= count;                                                                                         \
 }                                                                                                                      \
-static inline T* T##_unsafe_at(Darray_##T* darray, unsigned int index)  {                                              \
+static inline T* T##_unsafe_at(Darray_##T* darray, size_t index) {                                                     \
    return &darray->node[index];                                                                                        \
 }                                                                                                                      \
-static inline T* T##_at(Darray_##T* darray, unsigned int index)  {  /* Beware of segfaults after calling this */       \
+static inline T* T##_at(Darray_##T* darray, size_t index)  {  /* Beware of segfaults after calling this */             \
    if (index >= darray->size) return NULL;                                                                             \
    return &darray->node[index];                                                                                        \
 }                                                                                                                      \
-static inline bool T##_reserve(Darray_##T* darray, unsigned int block)  {                                              \
+static inline bool T##_reserve(Darray_##T* darray, size_t block)  {                                                    \
    if (block <= darray->capacity) return true;                                                                         \
    T* tmp = realloc(darray->node, sizeof(T) * block);                                                                  \
    if (!tmp) return false;                                                                                             \
